Refactored the currency dir
